perf(build): unity release builds — std::math lands at C parity (SIMD campaign result)
The SIMD-for-std::math campaign's M0 measurement showed the math types are already
SIMD-friendly: when the ops INLINE, clang at -O3 auto-vectorizes them at parity with
C. What blocked it in shipped binaries was the build model — `./kama build` compiled
each module as a separate TU with no LTO, so std::math never inlined into user hot
loops and shipped as scalar out-of-line calls (native `math` bench: 5.3x C).

Fix: `--release` native builds now fold every module into ONE unity translation unit
(kama.driver.cpp routes the multi-unit + release && !wasm case through the existing
transpileProgramToSingleFile merge, extended to forward the -lm / link hints). kama
has no incremental object cache — a build already hands all .c to a single clang
invocation — so the multi-file split bought nothing and only cost cross-module
inlining. Debug keeps per-module .c for faithful stepping; wasm keeps its path.

Result (native `math`, aarch64 bench): kama 27.1ms -> 3.13ms, now dead-even with C
(3.09ms) and ahead of C++/Rust; 25-38x faster than Go/C#/Java. Measured 29->2ms on
the isolated loop. Unity beats LTO decisively — LTO across separate TUs recovers only
part of the gap (ThinLTO 11ms, full LTO 8ms) because its cross-module inliner is far
more conservative than a real single TU; all link in ~the same time. No SIMD emitter
code, no API/layout change, no vector_size primitive needed for the common case.

Correctness: debug suite 664/0; a --release/unity sweep of every native fixture is
green (only non-builds are the web-only net_ws/rtc/wt fixtures, which don't build
native in either mode). Also aligned the `math` bench workload to the suite
convention (accumulate in float, cast once) so it measures math throughput rather
than a per-iteration float->int cast artifact; checksum unchanged, fairness gate
green across all 11 languages. docs/design/simd.md § M0 + ROADMAP §2 + RESULTS.md
updated.

Residual SIMD-codegen work is optional and contained: the Quat Hamilton product
stays scalar even when inlined (a hand-SIMD candidate) and the f64/DVec family.

+++ bench/src/cpp/math.cpp
@@ -30,7 +30,7 @@ static V4 quat_mul(V4 a, V4 b){   // Hamilton product (x,y,z,w)
 
 int main(){
   M4 m = { {1,1,0,0}, {0,1,1,0}, {0,0,1,1}, {1,0,0,1} };
-  uint64_t sum = 0;
+  double sum = 0.0;
   for(uint64_t i=0;i<2000000ULL;i++){
     float s = (float)(i % 8);
     V4 a = { s,   s+1, s+2, s+3 };
@@ -50,7 +50,7 @@ int main(){
               + (mv.x + mv.y + mv.z + mv.w)
               + (mm.c0.x + mm.c1.y + mm.c2.z + mm.c3.w)
               + qdot;
-    sum += (uint64_t)acc;
+    sum += acc;
   }
-  return (int)(sum % 256);
+  return (int)((uint64_t)sum % 256);
 }
